Move headers and sources around for better org
