+++ server/modules/selva/common.mk
@@ -1,7 +1,7 @@
 # find the OS
 uname_S := $(shell sh -c 'uname -s 2>/dev/null || echo not')
 
-CFLAGS := -std=gnu99 -Wall -Wextra -Wno-unused-value -Wno-zero-length-array -Wno-c11-extensions
+CFLAGS := -std=gnu99 -Wall -Wextra -Wno-unused-value
 
 # Compile flags for linux / osx
 ifeq ($(uname_S),Linux) # Assume Intel x86-64 Linux
@@ -10,5 +10,6 @@ ifeq ($(uname_S),Linux) # Assume Intel x86-64 Linux
 	#-ftree-vectorizer-verbose=5 -fopt-info-vec-missed
 endif
 ifeq ($(uname_S),Darwin) # Assume x86-64 macOS
-	CFLAGS += -g -march=x86-64 -mtune=core-avx2 -mfpmath=sse -mavx -mavx2
+	CFLAGS += -g -march=x86-64 -mtune=core-avx2 -mfpmath=sse -mavx -mavx2 \
+			  -Wno-zero-length-array -Wno-c11-extensions
 endif
