+++ server/modules/selva/include/rms.h
@@ -1,8 +1,41 @@
 #pragma once
-#ifndef _RMS_COMPRESSOR_H_
-#define _RMS_COMPRESSOR_H_
+#ifndef _SELVA_RMS_H_
+#define _SELVA_RMS_H_
 
 struct RedisModuleString;
+
+/**
+ * @addtogroup RMS
+ * @{
+ */
+
+/**
+ * @addtogroup shared
+ * @{
+ */
+
+/**
+ * Share a RedisModuleString.
+ * Anything passed to this function is expected to be needed forever read-only
+ * and there is no way to free anything nor determine if something could be
+ * freed. This storage model is ideal for things like type strings that are
+ * unlikely to ever change or change extremely rarely but are still used
+ * everywhere.
+ *
+ * @returns If rms is not shared yet it will be added to the internal data structure;
+ *          If rms is shared a pointer to the previously shared RedisModuleString is returned;
+ *          A NULL pointer is returned if adding rms to the internal data structure fails.
+ */
+struct RedisModuleString *Share_RMS(const char *key_str, size_t key_len, struct RedisModuleString *rms);
+
+/**
+ * @}
+ */
+
+/**
+ * @addtogroup compressor
+ * @{
+ */
 
 /**
  * Compressed RedisModuleString.
@@ -42,4 +75,12 @@ int rms_compress(struct compressed_rms *out, RedisModuleString *in);
  */
 int rms_decompress(struct RedisModuleString **out, struct compressed_rms *in);
 
-#endif /* _RMS_COMPRESSOR_H_ */
+/**
+ * @}
+ */
+
+/**
+ * @}
+ */
+
+#endif /* _SELVA_RMS_H_ */
