+++ server/modules/selva/module/find_index/find_index.c
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2021-2022 SAULX
+ * SPDX-License-Identifier: (MIT WITH selva-exception) OR AGPL-3.0-only
+ */
 #include <limits.h>
 #include <stddef.h>
 #include <stdio.h>
