+++ server/modules/selva/test/hierarchy_detached-mock.c
@@ -1,7 +1,7 @@
 #include "redismodule.h"
 #include "errors.h"
 #include "hierarchy.h"
-#include "hierarchy_detached.h"
+#include "../module/hierarchy/hierarchy_detached.h"
 
 int SelvaHierarchyDetached_Get(struct SelvaHierarchy *hierarchy, const Selva_NodeId node_id, struct compressed_rms **compressed) {
     return SELVA_ENOENT;
