+++ server/modules/selva/test/units/test-edge.mk
@@ -1,18 +1,5 @@
 TEST_SRC += test-edge.c
 SRC-edge += ../redis-alloc.c ../hierarchy-utils.c ../find-index-mock.c ../redis-rdb.c ../rpn-mock.c ../subscriptions-mock.c ../errors-mock.c ../hierarchy_detached-mock.c ../rms_compressor-mock.c
-SRC-edge += ../../module/alias.c
-SRC-edge += ../../module/arg_parser.c
-SRC-edge += ../../module/comparator.c
-SRC-edge += ../../module/config.c
-SRC-edge += ../../module/edge.c
-SRC-edge += ../../module/edge_constraint.c
-SRC-edge += ../../module/errors.c
-SRC-edge += ../../module/hierarchy.c
-SRC-edge += ../../module/selva_node.c
-SRC-edge += ../../module/selva_object.c
-SRC-edge += ../../module/selva_set.c
-SRC-edge += ../../module/selva_type.c
-SRC-edge += ../../module/shared.c
 SRC-edge += ../../lib/rmutil/sds.c
 SRC-edge += ../../lib/util/auto_free.c
 SRC-edge += ../../lib/util/cstrings.c
@@ -21,3 +8,16 @@ SRC-edge += ../../lib/util/memrchr.c
 SRC-edge += ../../lib/util/strnstr.c
 SRC-edge += ../../lib/util/svector.c
 SRC-edge += ../../lib/util/trx.c
+SRC-edge += ../../module/alias.c
+SRC-edge += ../../module/arg_parser.c
+SRC-edge += ../../module/comparator.c
+SRC-edge += ../../module/config.c
+SRC-edge += ../../module/edge/edge.c
+SRC-edge += ../../module/edge/edge_constraint.c
+SRC-edge += ../../module/errors.c
+SRC-edge += ../../module/hierarchy/hierarchy.c
+SRC-edge += ../../module/rms/shared.c
+SRC-edge += ../../module/selva_node.c
+SRC-edge += ../../module/selva_object/selva_object.c
+SRC-edge += ../../module/selva_set.c
+SRC-edge += ../../module/selva_type.c
