+++ server/modules/selva/test/units/test-hierarchy.mk
@@ -1,14 +1,5 @@
 TEST_SRC += test-hierarchy.c
 SRC-hierarchy += ../redis-alloc.c ../hierarchy-utils.c ../find-index-mock.c ../redis-rdb.c ../rpn-mock.c ../edge-mock.c ../subscriptions-mock.c ../errors-mock.c ../hierarchy_detached-mock.c ../rms_compressor-mock.c
-SRC-hierarchy += ../../module/shared.c
-SRC-hierarchy += ../../module/alias.c
-SRC-hierarchy += ../../module/config.c
-SRC-hierarchy += ../../module/errors.c
-SRC-hierarchy += ../../module/hierarchy.c
-SRC-hierarchy += ../../module/selva_node.c
-SRC-hierarchy += ../../module/selva_object.c
-SRC-hierarchy += ../../module/selva_set.c
-SRC-hierarchy += ../../module/selva_type.c
 SRC-hierarchy += ../../lib/rmutil/sds.c
 SRC-hierarchy += ../../lib/util/auto_free.c
 SRC-hierarchy += ../../lib/util/cstrings.c
@@ -17,3 +8,12 @@ SRC-hierarchy += ../../lib/util/memrchr.c
 SRC-hierarchy += ../../lib/util/strnstr.c
 SRC-hierarchy += ../../lib/util/svector.c
 SRC-hierarchy += ../../lib/util/trx.c
+SRC-hierarchy += ../../module/alias.c
+SRC-hierarchy += ../../module/config.c
+SRC-hierarchy += ../../module/errors.c
+SRC-hierarchy += ../../module/hierarchy/hierarchy.c
+SRC-hierarchy += ../../module/rms/shared.c
+SRC-hierarchy += ../../module/selva_node.c
+SRC-hierarchy += ../../module/selva_object/selva_object.c
+SRC-hierarchy += ../../module/selva_set.c
+SRC-hierarchy += ../../module/selva_type.c
