+++ server/modules/selva/test/units/test-rpn.mk
@@ -1,17 +1,17 @@
 TEST_SRC += test-rpn.c
-SRC-rpn += ../subscriptions-mock.c
 SRC-rpn += ../../lib/rmutil/sds.c
 SRC-rpn += ../../lib/util/cstrings.c
 SRC-rpn += ../../lib/util/mempool.c
 SRC-rpn += ../../lib/util/memrchr.c
 SRC-rpn += ../../lib/util/strnstr.c
 SRC-rpn += ../../lib/util/svector.c
-SRC-rpn += ../../module/shared.c
-SRC-rpn += ../../module/rpn.c 
 SRC-rpn += ../../module/errors.c
-SRC-rpn += ../../module/selva_object.c 
-SRC-rpn += ../../module/selva_set.c 
+SRC-rpn += ../../module/rms/shared.c
+SRC-rpn += ../../module/rpn/rpn.c
+SRC-rpn += ../../module/selva_object/selva_object.c
+SRC-rpn += ../../module/selva_set.c
 SRC-rpn += ../../module/selva_type.c
 SRC-rpn += ../errors-mock.c
-SRC-rpn += ../redis-alloc.c
 SRC-rpn += ../hierarchy-mock.c
+SRC-rpn += ../redis-alloc.c
+SRC-rpn += ../subscriptions-mock.c
