+++ server/modules/selva/test/units/test-selva_object.mk
@@ -1,13 +1,13 @@
 TEST_SRC += test-selva_object.c
 SRC-selva_object += ../redis-alloc.c ../errors-mock.c
-SRC-selva_object += ../../module/errors.c
-SRC-selva_object += ../../module/shared.c
-SRC-selva_object += ../../module/selva_object.c
-SRC-selva_object += ../../module/selva_set.c
-SRC-selva_object += ../../module/selva_type.c
 SRC-selva_object += ../../lib/rmutil/sds.c
 SRC-selva_object += ../../lib/util/cstrings.c
 SRC-selva_object += ../../lib/util/mempool.c
 SRC-selva_object += ../../lib/util/memrchr.c
 SRC-selva_object += ../../lib/util/strnstr.c
 SRC-selva_object += ../../lib/util/svector.c
+SRC-selva_object += ../../module/errors.c
+SRC-selva_object += ../../module/rms/shared.c
+SRC-selva_object += ../../module/selva_object/selva_object.c
+SRC-selva_object += ../../module/selva_set.c
+SRC-selva_object += ../../module/selva_type.c
